Assign stoichiometry for reactions without carbon map in atomMapping_Reactions2_file_parser()
Some reactions may not have carbon map, because it is unknown, or may be just unneccessary.
A good example is the biomass equation. Though through the atomMapping_Reaction2_file_parser()
these reactions were not properly parsed. This commit fixes this, so the stoichiometry, of
reactions without a carbon is also parsed.